// src/components/media/imageSchema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ImageInstance {
  type: "image";
  etag: string;
  public_id: string;
  original_filename: string;
  filename: string;
  path: string;
  format: string;
  dimensions: {
    height: number;
    width: number;
    [k: string]: unknown;
  };
  bytes: number;
  url: string;
  secure_url: string;
  cld_url: string;
  cld_secure_url: string;
  tags: string[];
  alt: string;
  timestamp: string;
  modified: string;
  [k: string]: unknown;
}